+++ src/main/java/com/EIPplatform/model/dto/report/report05/airemmissionmanagement/airautomonitoringstat/AirAutoMonitoringStatDTO.java
@@ -3,7 +3,7 @@ package com.EIPplatform.model.dto.report.report05.airemmissionmanagement.airauto
 import lombok.*;
 import lombok.experimental.FieldDefaults;
 
-import java.math.BigDecimal;
+import java.lang.Double;
 
 @Data
 @Builder
@@ -22,8 +22,8 @@ public class AirAutoMonitoringStatDTO {
 
     Integer valError;
 
-    BigDecimal ratioReceivedDesign;
+    Double ratioReceivedDesign;
 
-    BigDecimal ratioErrorReceived;
+    Double ratioErrorReceived;
 
 }
