+++ src/main/java/com/EIPplatform/model/dto/report/report05/wastemanagement/recycleindustrialwastestat/RecycleIndustrialWasteStatDTO.java
@@ -3,7 +3,7 @@ package com.EIPplatform.model.dto.report.report05.wastemanagement.recycleindustr
 import lombok.*;
 import lombok.experimental.FieldDefaults;
 
-import java.math.BigDecimal;
+import java.lang.Double;
 
 @Data
 @NoArgsConstructor
@@ -13,7 +13,7 @@ import java.math.BigDecimal;
 public class RecycleIndustrialWasteStatDTO {
     Long recycleId;
     String transferOrg;
-    BigDecimal volumeCy;
+    Double volumeCy;
     String wasteTypeDesc;
-    BigDecimal volumePy;
+    Double volumePy;
 }
